- Modified the stack to store doubles instead of integers. This is required to achieve results like the following: > 1 1/2 / = 2

+++ src/main/java/AdditionOperator.java
@@ -11,10 +11,10 @@ public class AdditionOperator extends Operator {
    * {@link Integer#MAX_VALUE} or {@link Integer#MIN_VALUE} respectively will be returned instead.
    */
   @Override
-  public int apply(int operand1, int operand2) {
+  public double apply(double operand1, double operand2) {
     /* Cast to longs before adding in case of an overflow, then truncate the result before casting
      * back to an int if required. */
-    return truncateToBounds(((long) operand1) + ((long) operand2));
+    return truncateToBounds(operand1 + operand2);
   }
 
   @Override
